Sanitize sizes after upload | remove upload wrapper
